+++ index.html
@@ -164,6 +164,6 @@
       </div>
     </section>
   </template>
-
+<script src="js/main.js" type="module"></script>
 </body>
 </html>
